Moved test_gif_output to main.py and guiTest to gui -> also added an overall confirm button
